vvvv

+++ src/main/java/com/example/demoggggg/repository/ExerciseRepository.java
@@ -13,12 +13,12 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface ExerciseRepository extends JpaRepository<Exercise,Integer> {
-    Exercise findById(int id);
+public interface ExerciseRepository extends JpaRepository<Exercise,Long> {
+    Exercise findById(long id);
     @Transactional
     @Modifying
     @Query(value = "update exercise set name=:name,introduction=:introduction,content=:content,maxpoints=:maxpoints," +
             "correctsolution=:correctsolution," +
             "correctoutput=:correctoutput where id=:id", nativeQuery = true)
-    void updateById(@Param("id") int id,@Param("name") String name,@Param("introduction") String introduction,@Param("content") String content,@Param("maxpoints") int maxPoints,@Param("correctsolution") String correctSolution,@Param("correctoutput") String correctOutput);
+    void updateById(@Param("id") long id,@Param("name") String name,@Param("introduction") String introduction,@Param("content") String content,@Param("maxpoints") int maxPoints,@Param("correctsolution") String correctSolution,@Param("correctoutput") String correctOutput);
 }
